pyproject.toml
[tool.poetry]
name = "quizbot"
version = "0.1.0"
description = "A custom bot developed for Mr.Brown to quiz their server members before assigning a role and providing access to restricted channels."
authors = ["DLCHAMP <36091350+dlchamp@users.noreply.github.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
disnake = "^2.6.0"
python-dotenv = "^0.21.0"
loguru = "^0.6.0"
typing-extensions = "^4.3.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
